deduplicate _ignore_warnings
